# openapi.v1.yaml
openapi: 3.1.0
info:
  title: KJWebsite Backend API
  version: 1.0.0
  description: Public and admin API contract for KJWebsite microservices via API Gateway.
servers:
  - url: http://localhost:7000
    description: Local API Gateway
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login with email/password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Auth tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Revoke refresh token
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '204':
          description: Logged out
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Current user profile
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorObject:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid payload
      required: [code, message]
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorObject'
      required: [error]
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
      required: [email, password]
    RefreshTokenRequest:
      type: object
      properties:
        refresh_token:
          type: string
      required: [refresh_token]
    AuthTokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          example: 900
        token_type:
          type: string
          example: Bearer
        user:
          $ref: '#/components/schemas/UserProfile'
      required: [access_token, refresh_token, expires_in, token_type]
    UserProfile:
      type: object
      properties:
        id: { type: string }
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, editor]
      required: [id, email, role]